+++ src/backend/ci/core/auth/biz-auth-rbac/src/main/kotlin/com/tencent/devops/auth/service/RbacPermissionResourceGroupService.kt
@@ -75,8 +75,7 @@ class RbacPermissionResourceGroupService @Autowired constructor(
     private val permissionResourceService: PermissionResourceService,
     private val permissionGroupPoliciesService: PermissionGroupPoliciesService,
     private val dslContext: DSLContext,
-    private val authResourceGroupDao: AuthResourceGroupDao,
-    private val client: Client
+    private val authResourceGroupDao: AuthResourceGroupDao
 ) : PermissionResourceGroupService {
 
     companion object {
@@ -92,7 +91,6 @@ class RbacPermissionResourceGroupService @Autowired constructor(
         page: Int,
         pageSize: Int
     ): Pagination<IamGroupInfoVo> {
-        checkProjectApprovalStatus(resourceType, resourceCode)
         val resourceInfo = authResourceService.get(
             projectCode = projectId,
             resourceType = resourceType,
@@ -139,24 +137,6 @@ class RbacPermissionResourceGroupService @Autowired constructor(
             hasNext = iamGroupInfoVoList.size == pageSize,
             records = iamGroupInfoVoList
         )
-    }
-
-    private fun checkProjectApprovalStatus(resourceType: String, resourceCode: String) {
-        if (resourceType == AuthResourceType.PROJECT.value) {
-            val projectInfo =
-                client.get(ServiceProjectResource::class).get(resourceCode).data ?: throw ErrorCodeException(
-                    errorCode = ProjectMessageCode.PROJECT_NOT_EXIST,
-                    params = arrayOf(resourceCode)
-                )
-            val approvalStatus = ProjectApproveStatus.parse(projectInfo.approvalStatus)
-            if (approvalStatus.isCreatePending()) {
-                throw ErrorCodeException(
-                    errorCode = ProjectMessageCode.UNDER_APPROVAL_PROJECT,
-                    params = arrayOf(resourceCode),
-                    defaultMessage = "project {0} is being approved, please wait patiently, or contact the approver"
-                )
-            }
-        }
     }
 
     override fun listUserBelongGroup(
